pyproject.toml
[tool.poetry]
name = "mentors-app"
version = "0.1.0"
description = ""
authors = ["Dmitrii Chebotarev <skrubberqumo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.8"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
commitizen = "^2.20.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
